Data Sanitization (#49)
* Change UserTask sortValue to Int for floating-point safety

* Make current-or-overdue filter day-granular

* null-result checks on all join aggregations

---------

+++ task-services/src/main/java/com/taat/taskservices/model/UserTask.java
@@ -26,7 +26,7 @@ public class UserTask {
 
     private String taskId;
 
-    private Double sortValue;
+    private Integer sortValue;
 
     private Instant skipUntil;
